chore: use ts-morph and split up files
