+++ docs/src/scripts/extract-api/wrapper-extractors.ts
@@ -1,0 +1,4 @@
+import { extractAngularWrapperApi } from "./angular-wrapper-extractor";
+import { extractReactWrapperApi } from "./react-wrapper-extractor";
+
+export { extractAngularWrapperApi, extractReactWrapperApi };
