+++ package.json
@@ -115,6 +115,7 @@
     "svelte-check": "3.8.6",
     "swc-loader": "0.2.7",
     "ts-jest": "29.4.9",
+    "ts-morph": "^28.0.0",
     "ts-node": "10.9.2",
     "typescript": "5.9.3",
     "vite": "5.4.21",
